use random position array rather than shuffling the entire pixel data

+++ standalone/src/main.rs
@@ -85,14 +85,14 @@ fn main_loop() {
         let rng = SmallRng::from_entropy();
         let mut raytracer = Raytracer::new(width, height, rng);
 
-        let mut generator = RandomGenerator::new(height as i64, width as i64, &mut SmallRng::from_entropy());
+        let mut generator = RandomGenerator::new(height as usize, width as usize, &mut SmallRng::from_entropy());
         loop {
             //spp *= 2;
             raytracer.generate_pixel(&mut generator, scene.as_slice(), spp, &mut communicator);
             generator.next();
             while let Ok(received_command) = rx.try_recv() {
                 spp = 1;
-                generator.invalidate_pixels(width as i64, height as i64, &mut SmallRng::from_entropy());
+                generator.invalidate_pixels(width as usize, height as usize, &mut SmallRng::from_entropy());
                 communicator.invalidate_pixels();
                 // frame dependant is bad but it does the job.
                 raytracer.camera = match received_command {
